ai-companion/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Dependencies
openai==1.3.0
anthropic==0.7.0
transformers==4.35.0
torch==2.1.0
langchain==0.0.330
langchain-openai==0.0.2
sentence-transformers==2.2.2
numpy==1.24.3

# Voice Processing (Eleven Labs)
elevenlabs==0.2.27
pyaudio==0.2.11
soundfile==1.0.6
webrtcvad==2.0.10

# Database and Storage
sqlalchemy==2.0.23
sqlite3
redis==5.0.1
alembic==1.12.1

# Configuration and Logging
pyyaml==6.0.1
python-dotenv==1.0.0
structlog==23.2.0
rich==13.7.0

# Async and Networking
aiohttp==3.9.0
websockets==12.0

# Security
cryptography==41.0.7
bcrypt==4.1.2

# System Integration
psutil==5.9.6
watchdog==3.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Utility
click==8.1.7
tqdm==4.66.1
python-dateutil==2.8.2
